uw-26sp-cse-590a/590A-assignment1-main/requirements.txt
# JAX (TPU) — requires the extra index for libtpu wheels
--find-links https://storage.googleapis.com/jax-releases/libtpu_releases.html
jax[tpu]==0.8.1
jaxlib==0.8.1

# Training & checkpointing
orbax-checkpoint==0.11.28
optax @ git+https://github.com/google-deepmind/optax.git
grain

# Data
numpy
datasets
huggingface-hub==1.2.3
tiktoken
tqdm

# Config & logging
pyyaml
rich
